+++ backdrop/write/config/test.py
@@ -21,18 +21,18 @@ BUCKET_UPLOAD_FORMAT = {
 BUCKET_UPLOAD_FILTERS = {
     "evl_ceg_data": [
         "backdrop.core.upload.filters.first_sheet_filter",
-        "backdrop.contrib.evl_upload.ceg_volumes"
+        "backdrop.contrib.evl_upload_filters.ceg_volumes"
     ],
     "evl_services_volumetrics": [
         "backdrop.core.upload.filters.first_sheet_filter",
-        "backdrop.contrib.evl_upload.service_volumetrics"
+        "backdrop.contrib.evl_upload_filters.service_volumetrics"
     ],
     "evl_services_failures": [
-        "backdrop.contrib.evl_upload.service_failures"
+        "backdrop.contrib.evl_upload_filters.service_failures"
     ],
     "evl_channel_volumetrics": [
         "backdrop.core.upload.filters.first_sheet_filter",
-        "backdrop.contrib.evl_upload.channel_volumetrics"
+        "backdrop.contrib.evl_upload_filters.channel_volumetrics"
     ],
 }
